+++ Cargo.toml
@@ -27,28 +27,42 @@ tracing infrastructure before running tests.
 """
 include = ["src/lib.rs", "LICENSE-*", "README.md", "CHANGELOG.md"]
 
-[lib]
-proc-macro = true
-
 [[test]]
 name = "default_log_filter"
 required-features = ["log"]
 
 [features]
 default = ["log"]
-trace = []
-log = []
+# TODO: use "dep:{tracing-subscriber,evn_logger}" once our MSRV is 1.60 or higher.
+trace = ["tracing-subscriber", "test-log-macros/trace"]
+log = ["env_logger", "test-log-macros/log"]
+
+[workspace]
+members = ["macros"]
 
 [dependencies]
-proc-macro2 = {version = "1.0.32", default-features = false}
-quote = {version = "1.0", default-features = false}
-syn = {version = "2.0", default-features = false, features = ["full", "parsing", "printing", "proc-macro"]}
+test-log-macros = {path = "macros"}
+tracing-subscriber = {version = "0.3.17", default-features = false, optional = true, features = ["env-filter", "fmt"]}
+env_logger = {version = "0.9.3", default-features = false, optional = true}
 
 [dev-dependencies]
-env_logger = {version = "0.9", default-features = false}
 logging = {version = "0.4.8", package = "log"}
 test-case = {version = "3.1"}
 tokio = {version = "1.0", default-features = false, features = ["rt-multi-thread", "macros"]}
 tracing = {version = "0.1.20"}
-tracing-futures = {version = "0.2", default-features = false, features = ["std-future"]}
-tracing-subscriber = {version = "0.3.10", default-features = false, features = ["env-filter", "fmt", "std"]}
+
+# error: cannot find macro `__lazy_static_internal` in this scope
+#   --> ~/.cargo/registry/src/github.com-1ecc6299db9ec823/sharded-slab-0.1.4/src/tid.rs:33:1
+#    |
+# 33 | / lazy_static! {
+# 34 | |     static ref REGISTRY: Registry = Registry {
+# 35 | |         next: AtomicUsize::new(0),
+# 36 | |         free: Mutex::new(VecDeque::new()),
+# 37 | |     };
+# 38 | | }
+#    | |_^
+#    |
+#    = note: this error originates in the macro `lazy_static` (in Nightly builds, run with -Z macro-backtrace for more info)
+#
+# Probably fixed by https://github.com/rust-lang-nursery/lazy-static.rs/pull/107.
+_lazy_static_unused = { package = "lazy_static", version = "1.0.2" }
