+++ macros/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "test-log-macros"
+version = "0.2.13"
+authors = ["Daniel Mueller <deso@posteo.net>"]
+edition = "2018"
+rust-version = "1.56"
+license = "Apache-2.0 OR MIT"
+
+[lib]
+proc-macro = true
+
+[features]
+trace = []
+log = []
+
+[dependencies]
+proc-macro2 = {version = "1.0.32", default-features = false}
+quote = {version = "1.0", default-features = false}
+syn = {version = "2.0", default-features = false, features = ["full", "parsing", "printing", "proc-macro"]}
